add validation for link controller

+++ app/Controllers/Api/V1/LinkController.php
@@ -3,7 +3,6 @@
 namespace App\Controllers\Api\V1;
 
 use App\Controllers\Controller;
-use App\Services\DomainService;
 use App\Services\LinkService;
 use Exception;
 
@@ -20,22 +19,15 @@ class LinkController extends Controller
 
     public function short(): void
     {
-        $baseUrl = base_url();
-        $domainId = input('domain_id');
-        if ($domainId) {
-            $domain = (new DomainService())->findOrFail($domainId);
+        $this->validator->validate(input()->all(), [
+            'original' => 'required|url',
+            'domain_id' => 'nullable|numeric'
+        ]);
 
-            if (!$domain['active']) {
-                throw new Exception('Domain is not active');
-            }
-
-            $baseUrl = 'https://' . $domain['name'] . '/';
-        }
-
-        $entity = $this->linkService->short(input('original'), $this->userId, $domainId);
+        $short = $this->linkService->short(input('original'), $this->userId, input('domain_id'));
 
         response()->httpCode(201)->json([
-            'short_link' => $baseUrl . $entity->short
+            'short_link' => $short
         ]);
     }
 
@@ -49,6 +41,11 @@ class LinkController extends Controller
     public function delete(string $short): void
     {
         $link = $this->linkService->findOrFail($short);
+
+        if ($link['user_id'] and $link['user_id'] != $this->userId) {
+            throw new Exception('Access denied');
+        }
+
         $result = $this->linkService->delete($link['id']);
 
         response()->json([
@@ -67,7 +64,18 @@ class LinkController extends Controller
 
     public function update(string $short): void
     {
+        $this->validator->validate(input()->all(), [
+            'short' => 'nullable',
+            'original' => 'nullable|url',
+            'domain_id' => 'nullable|numeric'
+        ]);
+
         $link = $this->linkService->findOrFail($short);
+
+        if ($link['user_id'] and $link['user_id'] != $this->userId) {
+            throw new Exception('Access denied');
+        }
+
         $result = $this->linkService->update(
             $link['id'],
             input('short'),
